[#1379] neofs-cli: Move accounting command to a separate package
